Actualize version of dalapi
