+++ source/elements/oneDAL/dalapi/directives.py
@@ -62,14 +62,13 @@ class DoxyDirective(MacroDirective):
             x.add_doc(desc_str, level=level)
 
     def add_function_base(self, func, x: RstBuilder, is_free=True, level=0):
-        namespace = func.parent_fully_qualified_name if is_free else None
-        x.add_function(func.declaration, namespace, level=level)
         if func.doc and func.doc.description:
+            namespace = func.parent_fully_qualified_name if is_free else None
+            x.add_function(func.declaration, namespace, level=level)
             self.add_description(func.doc.description, x, level=level + 1)
-        self.add_params('tparam', func.template_parameters, x, level=level + 1)
-        self.add_params('param', func.parameters, x, level=level + 1)
-        x.add_blank_line()
-        if func.doc:
+            self.add_params('tparam', func.template_parameters, x, level=level + 1)
+            self.add_params('param', func.parameters, x, level=level + 1)
+            x.add_blank_line()
             self.add_preconditions(func.doc.preconditions, x, level=level + 1)
             self.add_postconditions(func.doc.postconditions, x, level=level + 1)
 
@@ -113,7 +112,8 @@ class ClassDirective(DoxyDirective):
 
     def rst(self, x: RstBuilder):
         class_ = self.ctx.index.find(self.arguments[0])
-        self.add_listing(class_, x)
+        if self.ctx.listing_enabled:
+            self.add_listing(class_, x)
         sphinx_class_decl = (f'{class_.template_declaration} {class_.name}'
                              if class_.template_declaration else class_.name)
         x.add_class(class_.kind, sphinx_class_decl,
@@ -165,6 +165,10 @@ class ClassDirective(DoxyDirective):
                               level=1)
         if property_def.doc and property_def.doc.description:
             desc = self.format_description(property_def.doc.description)
+            if property_def.default:
+                if not desc.endswith('.'):
+                    desc += '.'
+                desc += f' **Default value**: {property_def.default}'
             if desc:
                 x.add_doc(desc, level=2)
             else:
@@ -189,8 +193,8 @@ class FunctionDirective(DoxyDirective):
 
     def rst(self, x: RstBuilder):
         func = self.ctx.index.find(self.arguments[0])
-        # TODO: Add option to include listing
-        # self.add_listing(func, x)
+        if self.ctx.listing_enabled:
+            self.add_listing(func, x)
         self.add_function_base(func, x, is_free=True)
 
 @directive
@@ -203,7 +207,8 @@ class EnumClassDirective(DoxyDirective):
         enum = self.ctx.index.find(self.arguments[0])
         namespace = enum.parent_fully_qualified_name
 
-        self.add_listing(enum, x)
+        if self.ctx.listing_enabled:
+            self.add_listing(enum, x)
         x.add_blank_line()
         x.add_enumclass(enum.name, namespace)
 
@@ -235,7 +240,8 @@ class TagsNamespaceDirective(DoxyDirective):
     def rst(self, x: RstBuilder):
         methods_namespace = self.arguments[0]
         method_ns = self.ctx.index.find(methods_namespace)
-        self.add_listing(method_ns, x, remove_empty_lines=True)
+        if self.ctx.listing_enabled:
+            self.add_listing(method_ns, x, remove_empty_lines=True)
         self._add_classes(method_ns, x)
         self._add_typedefs(method_ns, x)
