+++ source/elements/oneDAL/dalapi/doxypy/__init__.py
@@ -3,6 +3,6 @@
 # SPDX-License-Identifier: MIT
 
 from .index import Index, index, to_dict, to_json, to_yaml
-from .loader import TransformerPass
+from .loader import TransformerPass, NameTransformer
 from .listing import ListingReader
 from . import model
