+++ source/elements/oneDAL/make.bat
@@ -1,9 +1,9 @@
+@echo off
 REM SPDX-FileCopyrightText: 2019-2020 Intel Corporation
 REM
 REM SPDX-License-Identifier: CC-BY-4.0
 REM SPDX-License-Identifier: MIT
 
-@echo off
 if /I %1 == html goto :html
 if /I %1 == pdf goto :pdf
 if /I %1 == doxygen goto :doxygen
